matchmaking/requirements.txt
Flask==2.0.2
PyJWT==2.3.0
requests==2.22.0
